archetypen_gpt/requirements.txt
openai>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0
tqdm>=4.65.0
asyncio>=3.4.3
aiofiles>=23.2.0
markdown>=3.5.0
jinja2>=3.1.0
colorama>=0.4.6
rich>=13.0.0
questionary>=2.0.0 